Small additions to Authenticator tests and organizing TestUtils

+++ tests/AuthenticatorTests.py
@@ -5,13 +5,7 @@ sys.path.append('../lib')
 from Authenticator import *
 from Config import *
 import os
-
-def InitConfigFile(filename, content):
-  if os.path.exists(filename): os.remove(filename)
-
-  f = open(filename, "w+")
-  f.write(content)
-  f.close()
+from TestUtils import *
 
 class TestAuthenticator(unittest.TestCase):
   def test_authentication(self):
@@ -20,7 +14,7 @@ class TestAuthenticator(unittest.TestCase):
   "usersdbname": "testusers.db"
 }
 """
-    os.remove("testusers.db")
+    if os.path.exists("testusers.db"): os.remove("testusers.db")
 
     InitConfigFile("authconfig.json", simplejson)
     config = SimpleConfig("authconfig.json")
@@ -33,6 +27,7 @@ class TestAuthenticator(unittest.TestCase):
     self.assertFalse(auth.Authenticate("david", "pepa"))
 
     auth2 = SimpleAuthenticator(config)
+    self.assertTrue(auth2.Registered("david"))
     self.assertTrue(auth2.Registered("pepe"))
     self.assertFalse(auth2.Registered("lluisa"))
     self.assertTrue(auth2.Authenticate("pepe", "pepa"))
